Disconnected product components from Redux
Componets that deal with products are
now reliant on context 100%
