+++ mission-control-client/src/components/dashboard/admin-dashboard/DashboardProject.js
@@ -10,7 +10,7 @@ import {ProductContext} from '../../../context/ProductContext'
 
 const DashboardProject = props => {
   // Context
-  const { productState } = useContext(ProductContext)
+  const { productState, setSelectedProject } = useContext(ProductContext)
 
   let allowDelete = true;
   if (props.projects) {
@@ -24,9 +24,11 @@ const DashboardProject = props => {
   const history = useHistory();
 
   // references setting active project on second div tag of return below
-  const handleClick = () => {
-    props.setActiveProject(props.el.id);
+  const handleClick = async () => {
+    // props.setActiveProject(props.el.id);
+    await setSelectedProject(props.el.id)
     history.push(`/admin/dashboard/${props.el.id}`);
+    console.log("FROM HANDLE CLICK", props.el.id)
   };
 
   const [DeleteState, executeDeleteMutation] = useMutation(deleteProject);
@@ -118,12 +120,4 @@ const DashboardProject = props => {
   );
 };
 
-const mapStateToProps = state => {
-  return {
-    activeProductStore: state.activeProductStore
-  };
-};
-
-export default connect((mapStateToProps), {
-  setActiveProject
-})(DashboardProject);
+export default DashboardProject
